requirements.txt
streamlit
numpy
pandas
plotly
tensorflow-cpu   # lighter than full tensorflow, works on cloud